+++ src/client/logs.v
@@ -1,13 +1,14 @@
 module client
 
-import db { BuildLog }
+import models { BuildLog, BuildLogFilter }
 import net.http { Method }
 import response { Response }
 import time
 
 // get_build_logs returns all build logs.
-pub fn (c &Client) get_build_logs() ?Response<[]BuildLog> {
-	data := c.send_request<[]BuildLog>(Method.get, '/api/logs', {})?
+pub fn (c &Client) get_build_logs(filter BuildLogFilter) ?Response<[]BuildLog> {
+	params := models.params_from(filter)
+	data := c.send_request<[]BuildLog>(Method.get, '/api/logs', params)?
 
 	return data
 }
@@ -41,8 +42,8 @@ pub fn (c &Client) get_build_log_content(id int) ?string {
 pub fn (c &Client) add_build_log(repo_id int, start_time time.Time, end_time time.Time, arch string, exit_code int, content string) ?Response<string> {
 	params := {
 		'repo':      repo_id.str()
-		'startTime': start_time.str()
-		'endTime':   end_time.str()
+		'startTime': start_time.unix_time().str()
+		'endTime':   end_time.unix_time().str()
 		'arch':      arch
 		'exitCode':  exit_code.str()
 	}