+++ src/console/git/git.v
@@ -5,6 +5,7 @@ import env
 import cron.expression { parse_expression }
 import client
 import console
+import models { GitRepoFilter }
 
 struct Config {
 	address    string [required]
@@ -21,11 +22,45 @@ pub fn cmd() cli.Command {
 			cli.Command{
 				name: 'list'
 				description: 'List the current repos.'
+				flags: [
+					cli.Flag{
+						name: 'limit'
+						description: 'How many results to return.'
+						flag: cli.FlagType.int
+					},
+					cli.Flag{
+						name: 'offset'
+						description: 'Minimum index to return.'
+						flag: cli.FlagType.int
+					},
+					cli.Flag{
+						name: 'repo'
+						description: 'Only return Git repos that publish to this repo.'
+						flag: cli.FlagType.string
+					},
+				]
 				execute: fn (cmd cli.Command) ? {
 					config_file := cmd.flags.get_string('config-file')?
 					conf := env.load<Config>(config_file)?
 
-					list(conf)?
+					mut filter := GitRepoFilter{}
+
+					limit := cmd.flags.get_int('limit')?
+					if limit != 0 {
+						filter.limit = u64(limit)
+					}
+
+					offset := cmd.flags.get_int('offset')?
+					if offset != 0 {
+						filter.offset = u64(offset)
+					}
+
+					repo := cmd.flags.get_string('repo')?
+					if repo != '' {
+						filter.repo = repo
+					}
+
+					list(conf, filter)?
 				}
 			},
 			cli.Command{
@@ -133,9 +168,9 @@ pub fn cmd() cli.Command {
 // ID. If multiple or none are found, an error is raised.
 
 // list prints out a list of all repositories.
-fn list(conf Config) ? {
+fn list(conf Config, filter GitRepoFilter) ? {
 	c := client.new(conf.address, conf.api_key)
-	repos := c.get_git_repos()?
+	repos := c.get_git_repos(filter)?
 	data := repos.map([it.id.str(), it.url, it.branch, it.repo])
 
 	println(console.pretty_table(['id', 'url', 'branch', 'repo'], data)?)