+++ src/server/logs.v
@@ -8,6 +8,7 @@ import db
 import time
 import os
 import util
+import models { BuildLog, BuildLogFilter }
 
 // get_logs returns all build logs in the database. A 'repo' query param can
 // optionally be added to limit the list of build logs to that repository.
@@ -17,11 +18,10 @@ fn (mut app App) get_logs() web.Result {
 		return app.json(http.Status.unauthorized, new_response('Unauthorized.'))
 	}
 
-	logs := if 'repo' in app.query {
-		app.db.get_build_logs_for_repo(app.query['repo'].int())
-	} else {
-		app.db.get_build_logs()
+	filter := models.from_params<BuildLogFilter>(app.query) or {
+		return app.json(http.Status.bad_request, new_response('Invalid query parameters.'))
 	}
+	logs := app.db.get_build_logs(filter)
 
 	return app.json(http.Status.ok, new_data_response(logs))
 }
@@ -70,13 +70,19 @@ fn (mut app App) post_log() web.Result {
 	}
 
 	// Parse query params
-	start_time := parse_query_time(app.query['startTime']) or {
+	start_time_int := app.query['startTime'].int()
+
+	if start_time_int == 0 {
 		return app.json(http.Status.bad_request, new_response('Invalid or missing start time.'))
 	}
+	start_time := time.unix(start_time_int)
 
-	end_time := parse_query_time(app.query['endTime']) or {
+	end_time_int := app.query['endTime'].int()
+
+	if end_time_int == 0 {
 		return app.json(http.Status.bad_request, new_response('Invalid or missing end time.'))
 	}
+	end_time := time.unix(end_time_int)
 
 	if 'exitCode' !in app.query {
 		return app.json(http.Status.bad_request, new_response('Missing exit code.'))
@@ -97,7 +103,7 @@ fn (mut app App) post_log() web.Result {
 	}
 
 	// Store log in db
-	log := db.BuildLog{
+	log := BuildLog{
 		repo_id: repo_id
 		start_time: start_time
 		end_time: end_time